pyproject.toml
[project]
name = "delta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "attrs>=25.3.0",
    "bokeh>=3.7.0",
    "duckdb>=1.2.2",
    "fastapi>=0.115.11",
    "gradio>=5.22.0",
    "huggingface-hub>=0.30.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "scrapy>=2.12.0",
    "scrapy-playwright>=0.0.43",
    "scrapy-user-agents>=0.1.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "vega-datasets>=0.9.0",
]
